Front-End-staging-main/annotation/back-end-staging-main/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
arabic-reshaper==3.0.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
autopep8==2.3.2
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
decorator==4.4.2
Django==4.2.17
django-cors-headers==4.6.0
django-debug-toolbar==5.0.1
django-debug-toolbar-template-profiler==2.1.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
fpdf==1.7.2
frozenlist==1.5.0
gprof2dot==2024.6.6
graphviz==0.20.3
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==1.0.3
multidict==6.1.0
numpy==2.0.2
opencv-python==4.11.0.86
pillow==11.1.0
proglog==0.1.12
propcache==0.3.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydub==0.25.1
PyJWT==2.10.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wrapt==1.17.2
yarl==1.18.3
gunicorn
